submissions/vinicius-cury/solution/analysis/requirements.txt
jupyter
papermill
nbformat
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn
supabase
psycopg2-binary
python-dotenv
anthropic
google-generativeai
ipykernel
